requirements.txt
# LangChain and LangGraph - Core packages
langchain==1.0.3
langchain-openai==1.0.1
langgraph==1.0.2

# OpenAI API
openai==2.6.1

# Configuration and utilities
pyyaml==6.0.3
python-dotenv==1.2.1

# Data handling
pydantic==2.12.3
